requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dlib @ file:///C:/Users/Jayesh%20Thatte/Documents/PERSONAL/Projects/College/sih-2023/dlib-19.24.99-cp312-cp312-win_amd64.whl#sha256=20c62e606ca4c9961305f7be3d03990380d3e6c17f8d27798996e97a73271862
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==1.26.4
opencv-python==4.11.0.86
pillow==11.1.0
pyfingerprint==1.5
pyserial==3.5
pyzbar==0.1.9
requests==2.32.3
urllib3==2.3.0
Werkzeug==3.1.3
